better component names, prettified user-list and removed comments
