+++ src/app.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Switch, Redirect } from 'react-router';
 import './app.css';
 import LoginScreen from './components/login-screen/login-screen';
-import UserList from './components/user-list/user-list';
+import UserListScreen from './components/user-list/user-list-screen';
 import { getAuthToken } from './utils';
 
 const App: React.FC = () => {
@@ -10,7 +10,7 @@ const App: React.FC = () => {
     <div className='app'>
       <Switch>
         <Route path='/login' component={LoginScreen} />
-        <Route path='/user-list'  component={UserList} />
+        <Route path='/user-list'  component={UserListScreen} />
         <Route
           path='/'
           render={() => {
