+++ src/index.tsx
@@ -12,9 +12,7 @@ const httpLink = createHttpLink({
 });
 
 const authLink = setContext((_, { headers }) => {
-  // get the authentication token from local storage if it exists
   const token = getAuthToken();
-  // return the headers to the context so httpLink can read them
   return {
     headers: {
       ...headers,
